rug::complex

Struct BorrowComplex

pub struct BorrowComplex<'a> { /* private fields */ }

Used to get a reference to a Complex number.

The struct implements Deref<Target = Complex>.

No memory is unallocated when this struct is dropped.

Examples

use rug::complex::BorrowComplex;
use rug::Complex;
let c = Complex::with_val(53, (4.2, -2.3));
let neg: BorrowComplex = c.as_neg();
// c is still valid
assert_eq!(c, (4.2, -2.3));
assert_eq!(*neg, (-4.2, 2.3));

Implementations

impl<'a> BorrowComplex<'a>

pub const unsafe fn from_raw(raw: mpc_t) -> BorrowComplex<'a>

Create a borrow from a raw MPC complex number.

Safety

The value must be initialized as a valid mpc_t.

The mpc_t type can be considered as a kind of pointer, so there can be multiple copies of it. BorrowComplex cannot mutate the value, so there can be other copies, but none of them are allowed to mutate the value.

The lifetime is obtained from the return type. The user must ensure the value remains valid for the duration of the lifetime.

Examples

use rug::complex::BorrowComplex;
use rug::Complex;
let c = Complex::with_val(53, (4.2, -2.3));
// Safety: c.as_raw() is a valid pointer.
let raw = unsafe { *c.as_raw() };
// Safety: c is still valid when borrow is used.
let borrow = unsafe { BorrowComplex::from_raw(raw) };
assert_eq!(c, *borrow);

pub const fn const_deref<'b>(b: &'b BorrowComplex<'a>) -> &'b Complex

Gets a reference to Complex from a BorrowComplex.

This is equivalent to taking the reference of the dereferencing operator * or to Deref::deref, but can also be used in constant context. Unless required in constant context, use the operator or trait instead.

Planned deprecation

This method will be deprecated when the unary * operator and the Deref trait are usable in constant context.

Examples

use core::ops::Deref;
use core::ptr;
use rug::complex::BorrowComplex;
use rug::Complex;

let c = Complex::with_val(53, (23.5, -12.25));
let b = c.as_conj();
let using_method: &Complex = BorrowComplex::const_deref(&b);
let using_operator: &Complex = &*b;
let using_trait: &Complex = b.deref();
assert_eq!(*using_method, (23.5, 12.25));
assert!(ptr::eq(using_method, using_operator));
assert!(ptr::eq(using_method, using_trait));

This method can be used to create a constant reference.

use rug::complex::{BorrowComplex, MiniComplex};
use rug::float::MiniFloat;
use rug::Complex;

const TWO_FLOAT: MiniFloat = MiniFloat::const_from_i8(2);
const THOUSAND_FLOAT: MiniFloat = MiniFloat::const_from_i16(1000);
const TWO_1000I_MINI: MiniComplex =
    MiniComplex::const_from_parts(TWO_FLOAT, THOUSAND_FLOAT);
const TWO_1000I_BORROW: BorrowComplex = TWO_1000I_MINI.borrow();
const TWO_1000I: &Complex = BorrowComplex::const_deref(&TWO_1000I_BORROW);
assert_eq!(*TWO_1000I, (2, 1000));

If the constant cannot be initialized using MiniComplex, we can set up the limbs directly, though it is more complicated.

use core::ptr::NonNull;
use gmp_mpfr_sys::gmp::limb_t;
use gmp_mpfr_sys::mpfr::{mpfr_t, prec_t};
use gmp_mpfr_sys::mpc::mpc_t;
use rug::complex::BorrowComplex;
use rug::{Complex, Float};

const LIMBS: [limb_t; 2] = [5, 1 << (limb_t::BITS - 1)];
const LIMBS_PTR: *const [limb_t; 2] = &LIMBS;
const MANTISSA_DIGITS: u32 = limb_t::BITS * 2;
const MPC: mpc_t = mpc_t {
    re: mpfr_t {
        prec: MANTISSA_DIGITS as prec_t,
        sign: -1,
        exp: 1,
        d: unsafe { NonNull::new_unchecked(LIMBS_PTR.cast_mut().cast()) },
    },
    im: mpfr_t {
        prec: MANTISSA_DIGITS as prec_t,
        sign: 1,
        exp: 1,
        d: unsafe { NonNull::new_unchecked(LIMBS_PTR.cast_mut().cast()) },
    },
};
// Safety: MPFR will remain valid, and will not be changed.
const BORROW: BorrowComplex = unsafe { BorrowComplex::from_raw(MPC) };
const C: &Complex = BorrowComplex::const_deref(&BORROW);
let lsig = Float::with_val(MANTISSA_DIGITS, 5) >> (MANTISSA_DIGITS - 1);
let msig = 1u32;
let val = lsig + msig;
let check = Complex::from((-val.clone(), val));
assert_eq!(*C, check);