Struct rug::complex::SmallComplex

pub struct SmallComplex { /* fields omitted */ }

A small complex number that does not require any memory allocation.

This can be useful when you have real and imaginary numbers that are primitive integers or floats and you need a reference to a Complex.

The SmallComplex will have a precision according to the types of the primitives used to set its real and imaginary parts. Note that if different types are used to set the parts, the parts can have different precisions.

i8, u8: the part will have eight bits of precision.

i16, u16: the part will have 16 bits of precision.

i32, u32: the part will have 32 bits of precision.

i64, u64: the part will have 64 bits of precision.

i128, u128: the part will have 128 bits of precision.

isize, usize: the part will have 32 or 64 bits of precision, depending on the platform.

f32: the part will have 24 bits of precision.

f64: the part will have 53 bits of precision.

Special: the part will have the minimum possible precision.

The SmallComplex type can be coerced to a Complex, as it implements Deref<Target = Complex>.

Examples

use rug::{complex::SmallComplex, Complex};
// `a` requires a heap allocation
let mut a = Complex::with_val(53, (1, 2));
// `b` can reside on the stack
let b = SmallComplex::from((-10f64, -20.5f64));
a += &*b;
assert_eq!(*a.real(), -9);
assert_eq!(*a.imag(), -18.5);

Methods

impl SmallComplex

pub const fn new() -> Self

Creates a SmallComplex with value 0 and the minimum possible precision.

Examples

use rug::complex::SmallComplex;
let c = SmallComplex::new();
// Borrow c as if it were Complex.
assert_eq!(*c, 0);

pub unsafe fn as_nonreallocating_complex(&mut self) -> &mut Complex

Returns a mutable reference to a Complex number for simple operations that do not need to change the precision of the real or imaginary part.

Safety

It is undefined behaviour to modify the precision of the referenced Complex number or to swap it with another number.

Examples

use rug::complex::SmallComplex;
let mut c = SmallComplex::from((1.0f32, 3.0f32));
// rotation does not change the precision
unsafe {
    c.as_nonreallocating_complex().mul_i_mut(false);
}
assert_eq!(*c, (-3.0, 1.0));