Module rug::float

Multi-precision floating-point numbers with correct rounding.

This module provides support for floating-point numbers of type Float.

Structs

OrdFloat
A float that supports total ordering and hashing.

ParseFloatError
An error which can be returned when parsing a Float.

SmallFloat
A small float that does not require any memory allocation.

Enums

Constant
The available floating-point constants.

FreeCache
Specifies which cache to free.

Round
The rounding methods for floating-point values.

Special
Special floating-point values.

Traits

ToSmall
Types implementing this trait can be converted to SmallFloat.

Functions

allowed_exp_range
Returns the maximum allowed range for the exponent.

exp_max
Returns the maximum value for the exponent.

exp_min
Returns the minimum value for the exponent.

free_cache
Frees various caches and memory pools that are used internally.

prec_max
Returns the maximum value for the precision.

prec_min
Returns the minimum value for the precision.