rug::float

Enum Round

#[non_exhaustive]
pub enum Round {
    Nearest,
    Zero,
    Up,
    Down,
    AwayZero,
}

The rounding methods for floating-point values.

When rounding to the nearest, if the number to be rounded is exactly between two representable numbers, it is rounded to the even one, that is, the one with the least significant bit set to zero.

Examples

use rug::float::Round;
use rug::ops::AssignRound;
use rug::Float;
let mut f4 = Float::new(4);
f4.assign_round(10.4, Round::Nearest);
assert_eq!(f4, 10);
f4.assign_round(10.6, Round::Nearest);
assert_eq!(f4, 11);
f4.assign_round(-10.7, Round::Zero);
assert_eq!(f4, -10);
f4.assign_round(10.3, Round::Up);
assert_eq!(f4, 11);

Rounding to the nearest will round numbers exactly between two representable numbers to the even one.

use rug::float::Round;
use rug::ops::AssignRound;
use rug::Float;
// 24 is 11000 in binary
// 25 is 11001 in binary
// 26 is 11010 in binary
// 27 is 11011 in binary
// 28 is 11100 in binary
let mut f4 = Float::new(4);
f4.assign_round(25, Round::Nearest);
assert_eq!(f4, 24);
f4.assign_round(27, Round::Nearest);
assert_eq!(f4, 28);

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Nearest

Round towards the nearest, with ties rounding to even.

Zero

Round towards zero.

Up

Round towards plus infinity.

Down

Round towards minus infinity.

AwayZero

Round away from zero.

Implementations

impl Round

pub fn reverse(self) -> Round

Reverses the rounding direction.

Up becomes Down.

Down becomes Up.

Other variants are not affected.

Examples

use rug::float::Round;

assert_eq!(Round::Up.reverse(), Round::Down);
assert_eq!(Round::Down.reverse(), Round::Up);
assert_eq!(Round::Nearest.reverse(), Round::Nearest);