Struct rug::integer::ParseIntegerError

pub struct ParseIntegerError { /* fields omitted */ }

An error which can be returned when parsing an Integer.

See the Integer::parse_radix method for details on what strings are accepted.

Examples

use rug::{integer::ParseIntegerError, Integer};
// This string is not an integer.
let s = "something completely different (_!_!_)";
let error: ParseIntegerError = match Integer::parse_radix(s, 4) {
    Ok(_) => unreachable!(),
    Err(error) => error,
};
println!("Parse error: {}", error);