Module rug::ops

Operations on numbers.

See the documentation for each trait method to see a usage example.

Traits

AddAssignRound
Compound addition and assignment with a specified rounding method.

AddFrom
Compound addition and assignment to the rhs operand.

AddFromRound
Compound addition and assignment to the rhs operand with a specified rounding method.

AssignRound
Assignment with a specified rounding method.

BitAndFrom
Compound bitwise AND and assignment to the rhs operand.

BitOrFrom
Compound bitwise OR and assignment to the rhs operand.

BitXorFrom
Compound bitwise XOR and assignment to the rhs operand.

DivAssignRound
Compound division and assignment with a specified rounding method.

DivFrom
Compound division and assignment to the rhs operand.

DivFromRound
Compound division and assignment to the rhs operand with a specified rounding method.

DivRounding
Rounding variants of division.

DivRoundingAssign
Compound assignment and rounding variants of division.

DivRoundingFrom
Compound assignment to the rhs operand and rounding variants of division.

MulAssignRound
Compound multiplication and assignment with a specified rounding method.

MulFrom
Compound multiplication and assignment to the rhs operand.

MulFromRound
Compound multiplication and assignment to the rhs operand with a specified rounding method.

NegAssign
Compound negation and assignment.

NotAssign
Compound bitwise complement and assignement.

Pow
The power operation.

PowAssign
Compound power operation and assignment.

PowAssignRound
Compound power operation and assignment with a specified rounding method.

PowFrom
Compound power operation and assignment to the rhs operand.

PowFromRound
Compound power operation and assignment to the rhs operand with a specified rounding method.

RemFrom
Compound remainder operation and assignment to the rhs operand.

RemRounding
Rounding variants of the remainder operation.

RemRoundingAssign
Compound assignment and rounding variants of the remainder operation.

RemRoundingFrom
Compound assignment to the rhs operand and rounding variants of the remainder operation.

ShlFrom
Compound left shift and assignment to the rhs operand.

ShrFrom
Compound right shift and assignment to the rhs operand.

SubAssignRound
Compound subtraction and assignment with a specified rounding method.

SubFrom
Compound subtraction and assignment to the rhs operand.

SubFromRound
Compound subtraction and assignment to the rhs operand with a specified rounding method.