Trait rug::rand::MutRandState

pub trait MutRandState: SealedMutRandState { }

Used to pass the state of random number generators by mutable reference.

This trait is implemented by

RandState, which is thread safe and implements Send and Sync.

ThreadRandState, which can only be used in a single thread.

This trait is sealed and cannot be implemented for more types.