rug::rational

Struct BorrowRational

pub struct BorrowRational<'a> { /* private fields */ }

Used to get a reference to a Rational number.

The struct implements Deref<Target = Rational>.

No memory is unallocated when this struct is dropped.

Examples

use rug::rational::{BorrowRational, MiniRational};
use rug::Rational;
let r = Rational::from((42, 3));
let neg: BorrowRational = r.as_neg();
// r is still valid
assert_eq!(r, MiniRational::from((42, 3)));
assert_eq!(*neg, MiniRational::from((-42, 3)));

Implementations

impl<'a> BorrowRational<'a>

pub const unsafe fn from_raw(raw: mpq_t) -> BorrowRational<'a>

Create a borrow from a raw GMP rational number.

Safety

The value must be initialized as a valid mpq_t.

The mpq_t type can be considered as a kind of pointer, so there can be multiple copies of it. BorrowRational cannot mutate the value, so there can be other copies, but none of them are allowed to mutate the value.

The lifetime is obtained from the return type. The user must ensure the value remains valid for the duration of the lifetime.

The numerator and denominator must be in canonical form, as the rest of the library assumes that they are. Most GMP functions leave the rational number in canonical form, but assignment functions do not. Check the GMP documentation for details.

Examples

use rug::rational::BorrowRational;
use rug::Rational;
let r = Rational::from((42, 3));
// Safety: r.as_raw() is a valid pointer.
let raw = unsafe { *r.as_raw() };
// Safety: r is still valid when borrow is used.
let borrow = unsafe { BorrowRational::from_raw(raw) };
assert_eq!(r, *borrow);

pub const fn const_deref<'b>(b: &'b BorrowRational<'a>) -> &'b Rational

Gets a reference to Rational from a BorrowRational.

This is equivalent to taking the reference of the dereferencing operator * or to Deref::deref, but can also be used in constant context. Unless required in constant context, use the operator or trait instead.

Planned deprecation

This method will be deprecated when the unary * operator and the Deref trait are usable in constant context.

Examples

use core::ops::Deref;
use core::ptr;
use rug::rational::{BorrowRational, MiniRational};
use rug::Rational;

let i = Rational::from((-5, 7));
let b = i.as_recip();
let using_method: &Rational = BorrowRational::const_deref(&b);
let using_operator: &Rational = &*b;
let using_trait: &Rational = b.deref();
assert_eq!(*using_trait, MiniRational::from((-7, 5)));
assert!(ptr::eq(using_method, using_operator));
assert!(ptr::eq(using_method, using_trait));

This method can be used to create a constant reference.

use rug::integer::MiniInteger;
use rug::rational::{BorrowRational, MiniRational};
use rug::Rational;

const TWO_INT: MiniInteger = MiniInteger::const_from_i32(2);
const TWO_MINI: MiniRational = MiniRational::const_from_integer(TWO_INT);
const TWO_BORROW: BorrowRational = TWO_MINI.borrow();
const TWO: &Rational = BorrowRational::const_deref(&TWO_BORROW);
assert_eq!(*TWO, 2);

Creating a constant with MiniRational is rather limited. If the constant cannot be initialized that way, we can set up the limbs directly, though it is more complicated.

use gmp_mpfr_sys::gmp;
use gmp_mpfr_sys::gmp::{limb_t, mpq_t};
use rug::rational::BorrowRational;
use rug::{Integer, Rational};

const NUMER_LIMBS: [limb_t; 2] = [0, 5];
const DENOM_LIMBS: [limb_t; 1] = [3];
const MPQ: mpq_t = unsafe {
    mpq_t {
        num: gmp::MPZ_ROINIT_N(NUMER_LIMBS.as_ptr().cast_mut(), -2),
        den: gmp::MPZ_ROINIT_N(DENOM_LIMBS.as_ptr().cast_mut(), 1),
    }
};
// Safety: MPQ will remain valid, and will not be changed.
const BORROW: BorrowRational = unsafe { BorrowRational::from_raw(MPQ) };
const R: &Rational = BorrowRational::const_deref(&BORROW);
let numer_check =
    -((Integer::from(NUMER_LIMBS[1]) << gmp::NUMB_BITS) + NUMER_LIMBS[0]);
let denom_check = Integer::from(DENOM_LIMBS[0]);
let check = Rational::from((&numer_check, &denom_check));
assert_eq!(*R, check);
assert_eq!(*R.numer(), *check.numer());
assert_eq!(*R.denom(), *check.denom());