Struct rug::rational::SmallRational

#[repr(C)]

pub struct SmallRational { /* fields omitted */ }

A small rational number that does not require any memory allocation.

This can be useful when you have a numerator and denominator that are primitive integer-types such as i64 or u8, and you need a reference to a Rational.

Although no allocation is required, setting the value of a SmallRational does require some computation, as the numerator and denominator need to be canonicalized.

The SmallRational type can be coerced to a Rational, as it implements Deref<Target = Rational>.

Examples

use rug::Rational;
use rug::rational::SmallRational;
// `a` requires a heap allocation
let mut a = Rational::from((100, 13));
// `b` can reside on the stack
let b = SmallRational::from((-100, 21));
a /= &*b;
assert_eq!(*a.numer(), -21);
assert_eq!(*a.denom(), 13);

Methods

impl SmallRational

pub fn new() -> SmallRational

Creates a SmallRational with value 0.

Examples

use rug::rational::SmallRational;
let r = SmallRational::new();
// Use r as if it were Rational.
assert_eq!(*r.numer(), 0);
assert_eq!(*r.denom(), 1);

pub unsafe fn as_nonreallocating_rational(&mut self) -> &mut Rational

Returns a mutable reference to a Rational number for simple operations that do not need to allocate more space for the numerator or denominator.

Safety

It is undefined behaviour to perform operations that reallocate the internal data of the referenced Rational number or to swap it with another number.

Examples

use rug::rational::SmallRational;
let mut r = SmallRational::from((-15i32, 47i32));
let num_capacity = r.numer().capacity();
let den_capacity = r.denom().capacity();
// reciprocating this will not require reallocations
unsafe {
    r.as_nonreallocating_rational().recip_mut();
}
assert_eq!(*r, (-47, 15));
assert_eq!(r.numer().capacity(), num_capacity);
assert_eq!(r.denom().capacity(), den_capacity);

pub unsafe fn from_canonical<Num, Den>(num: Num, den: Den) -> SmallRational where
SmallInteger: Assign<Num> + Assign<Den>,

Creates a SmallRational from a numerator and denominator, assuming they are in canonical form.

Safety

This method leads to undefined behaviour if den is zero or if num and den have common factors.

Examples

use rug::rational::SmallRational;
let from_unsafe = unsafe {
    SmallRational::from_canonical(-13, 10)
};
// from_safe is canonicalized to the same form as from_unsafe
let from_safe = SmallRational::from((130, -100));
assert_eq!(from_unsafe.numer(), from_safe.numer());
assert_eq!(from_unsafe.denom(), from_safe.denom());

pub unsafe fn from_canonicalized_32(
neg: bool,
num: u32,
den: u32
) -> SmallRational

Deprecated since 0.9.2
: use from_canonical instead; for example SmallRational::from_cananoicalized_32(true, 13, 10) can be replaced with SmallRational::from_canonical(-13, 10).

Creates a SmallRational from a 32-bit numerator and denominator, assuming they are in canonical form.

pub unsafe fn from_canonicalized_64(
neg: bool,
num: u64,
den: u64
) -> SmallRational

Deprecated since 0.9.2
: use from_canonical instead; for example SmallRational::from_cananoicalized_64(true, 13, 10) can be replaced with SmallRational::from_canonical(-13, 10).

Creates a SmallRational from a 64-bit numerator and denominator, assuming they are in canonical form.

pub unsafe fn assign_canonicalized_32(&mut self, neg: bool, num: u32, den: u32)

Deprecated since 0.9.2
: use as_nonreallocating_rational and assign_canonical instead; for example r.assign_canonicalized_32(true, 13u32, 10u32) can be replaced with r.as_nonreallocating_rational().assign_canonical(-13i32, 10u32).

Sets a SmallRational to a 32-bit numerator and denominator, assuming they are in canonical form.

pub unsafe fn assign_canonicalized_64(&mut self, neg: bool, num: u64, den: u64)

Deprecated since 0.9.2
: use as_nonreallocating_rational and assign_canonical instead; for example r.assign_canonicalized_32(true, 13u64, 10u64) can be replaced with r.as_nonreallocating_rational().assign_canonical(-13i64, 10u64).

Sets a SmallRational to a 64-bit numerator and denominator, assuming they are in canonical form.