rug::complex

Struct SmallComplex

pub struct SmallComplex { /* private fields */ }

๐Deprecated since 1.23.0: use MiniComplex instead

A small complex number that did not require any memory allocation until version 1.23.0.

Because of a soundness issue, this has been deprecated and replaced by MiniComplex. To fix the soundness issue, this struct now uses allocations like Complex itself, so it is less efficient than MiniComplex.

The SmallComplex type can be coerced to a Complex, as it implements Deref<Target = Complex>.

Examples

#![allow(deprecated)]

use rug::complex::SmallComplex;
use rug::Complex;
// `a` requires a heap allocation
let mut a = Complex::with_val(53, (1, 2));
// `b` can reside on the stack
let b = SmallComplex::from((-10f64, -20.5f64));
a += &*b;
assert_eq!(*a.real(), -9);
assert_eq!(*a.imag(), -18.5);

Implementations

impl SmallComplex

pub const fn new() -> Self

Creates a SmallComplex with value 0 and the minimum possible precision.

Examples

#![allow(deprecated)]

use rug::complex::SmallComplex;
let c = SmallComplex::new();
// Borrow c as if it were Complex.
assert_eq!(*c, 0);

pub unsafe fn as_nonreallocating_complex(&mut self) -> &mut Complex

Returns a mutable reference to a Complex number for simple operations that do not need to change the precision of the real or imaginary part.

Safety

It is undefined behavior to modify the precision of the referenced Complex number or to swap it with another number.

Examples

#![allow(deprecated)]

use rug::complex::SmallComplex;
let mut c = SmallComplex::from((1.0f32, 3.0f32));
// rotation does not change the precision
unsafe {
    c.as_nonreallocating_complex().mul_i_mut(false);
}
assert_eq!(*c, (-3.0, 1.0));