rug::float

Struct SmallFloat

pub struct SmallFloat { /* private fields */ }

๐Deprecated since 1.23.0: use MiniFloat instead

A small float that did not require any memory allocation until version 1.23.0.

Because of a soundness issue, this has been deprecated and replaced by MiniFloat. To fix the soundness issue, this struct now uses allocations like Float itself, so it is less efficient than MiniFloat.

The SmallFloat type can be coerced to a Float, as it implements Deref<Target = Float>.

Examples

#![allow(deprecated)]

use rug::float::SmallFloat;
use rug::Float;
// `a` requires a heap allocation, has 53-bit precision
let mut a = Float::with_val(53, 250);
// `b` can reside on the stack
let b = SmallFloat::from(-100f64);
a += &*b;
assert_eq!(a, 150);
// another computation:
a *= &*b;
assert_eq!(a, -15000);

Implementations

impl SmallFloat

pub const fn new() -> Self

Creates a SmallFloat with value 0 and the minimum possible precision.

Examples

#![allow(deprecated)]

use rug::float::SmallFloat;
let f = SmallFloat::new();
// Borrow f as if it were Float.
assert_eq!(*f, 0);

pub unsafe fn as_nonreallocating_float(&mut self) -> &mut Float

Returns a mutable reference to a Float for simple operations that do not need to change the precision of the number.

Safety

It is undefined behavior modify the precision of the referenced Float or to swap it with another number.

Examples

#![allow(deprecated)]

use rug::float::SmallFloat;
let mut f = SmallFloat::from(1.0f32);
// addition does not change the precision
unsafe {
    *f.as_nonreallocating_float() += 2.0;
}
assert_eq!(*f, 3.0);