Struct rug::float::SmallFloat

pub struct SmallFloat { /* fields omitted */ }

A small float that does not require any memory allocation.

This can be useful when you have a primitive number type but need a reference to a Float. The SmallFloat will have a precision according to the type of the primitive used to set its value.

i8, u8: the SmallFloat will have eight bits of precision.

i16, u16: the SmallFloat will have 16 bits of precision.

i32, u32: the SmallFloat will have 32 bits of precision.

i64, u64: the SmallFloat will have 64 bits of precision.

i128, u128: the SmallFloat will have 128 bits of precision.

isize, usize: the SmallFloat will have 32 or 64 bits of precision, depending on the platform.

f32: the SmallFloat will have 24 bits of precision.

f64: the SmallFloat will have 53 bits of precision.

Special: the SmallFloat will have the minimum possible precision.

The SmallFloat type can be coerced to a Float, as it implements Deref<Target = Float>.

Examples

use rug::{float::SmallFloat, Float};
// `a` requires a heap allocation, has 53-bit precision
let mut a = Float::with_val(53, 250);
// `b` can reside on the stack
let b = SmallFloat::from(-100f64);
a += &*b;
assert_eq!(a, 150);
// another computation:
a *= &*b;
assert_eq!(a, -15000);

Implementations

impl SmallFloat

pub const fn new() -> Self

Creates a SmallFloat with value 0 and the minimum possible precision.

Examples

use rug::float::SmallFloat;
let f = SmallFloat::new();
// Borrow f as if it were Float.
assert_eq!(*f, 0);

pub unsafe fn as_nonreallocating_float(&mut self) -> &mut Float

Returns a mutable reference to a Float for simple operations that do not need to change the precision of the number.

Safety

It is undefined behavior modify the precision of the referenced Float or to swap it with another number.

Examples

use rug::float::SmallFloat;
let mut f = SmallFloat::from(1.0f32);
// addition does not change the precision
unsafe {
    *f.as_nonreallocating_float() += 2.0;
}
assert_eq!(*f, 3.0);