Trait rug::float::ToSmall

pub trait ToSmall: SealedToSmall { }

Types implementing this trait can be converted to SmallFloat.

The following are implemented when T implements ToSmall:

Assign<T> for SmallFloat

From<T> for SmallFloat

This trait is sealed and cannot be implemented for more types; it is implemented for the integer types i8, i16, i32, i64, i128, isize, u8, u16, u32, u64, u128 and usize, and for the floating-point types f32 and f64.