rug::integer

Struct SmallInteger

pub struct SmallInteger { /* private fields */ }

๐Deprecated since 1.23.0: use MiniInteger instead

A small integer that did not require any memory allocation until version 1.23.0.

Because of a soundness issue, this has been deprecated and replaced by MiniInteger. To fix the soundness issue, this struct now uses allocations like Integer itself, so it is less efficient than MiniInteger.

The SmallInteger type can be coerced to an Integer, as it implements Deref<Target = Integer>.

Examples

#![allow(deprecated)]

use rug::integer::SmallInteger;
use rug::Integer;
// `a` requires a heap allocation
let mut a = Integer::from(250);
// `b` can reside on the stack
let b = SmallInteger::from(-100);
a.lcm_mut(&b);
assert_eq!(a, 500);
// another computation:
a.lcm_mut(&SmallInteger::from(30));
assert_eq!(a, 1500);

Implementations

impl SmallInteger

pub const fn new() -> Self

Creates a SmallInteger with value 0.

Examples

#![allow(deprecated)]

use rug::integer::SmallInteger;
let i = SmallInteger::new();
// Borrow i as if it were Integer.
assert_eq!(*i, 0);

pub unsafe fn as_nonreallocating_integer(&mut self) -> &mut Integer

Returns a mutable reference to an Integer for simple operations that do not need to allocate more space for the number.

Safety

It is undefined behavior to perform operations that reallocate the internal data of the referenced Integer or to swap it with another number.

Some GMP functions swap the allocations of their target operands; calling such functions with the mutable reference returned by this method can lead to undefined behavior.

Examples

#![allow(deprecated)]

use rug::integer::SmallInteger;
use rug::Assign;
let mut i = SmallInteger::from(1u64);
let capacity = i.capacity();
// another u64 will not require a reallocation
unsafe {
    i.as_nonreallocating_integer().assign(2u64);
}
assert_eq!(*i, 2);
assert_eq!(i.capacity(), capacity);