Trait rug::integer::UnsignedPrimitive

pub trait UnsignedPrimitive: SealedUnsignedPrimitive { }

Conversions between Integer and a slice of digits of this type are provided.

For conversion from digits to Integer, see Integer::from_digits and Integer::assign_digits. For conversion from Integer to digits, see Integer::significant_digits, Integer::to_digits, and Integer::write_digits.

This trait is sealed and cannot be implemented for more types; it is implemented for bool, u8, u16, u32, u64, u128 and usize.