rug::rational

Struct SmallRational

pub struct SmallRational { /* private fields */ }

๐Deprecated since 1.23.0: use MiniRational instead

A small rational number that did not require any memory allocation until version 1.23.0.

Because of a soundness issue, this has been deprecated and replaced by MiniRational. To fix the soundness issue, this struct now uses allocations like Rational itself, so it is less efficient than MiniRational.

The SmallRational type can be coerced to a Rational, as it implements Deref<Target = Rational>.

Examples

#![allow(deprecated)]

use rug::rational::SmallRational;
use rug::Rational;
// `a` requires a heap allocation
let mut a = Rational::from((100, 13));
// `b` can reside on the stack
let b = SmallRational::from((-100, 21));
a /= &*b;
assert_eq!(*a.numer(), -21);
assert_eq!(*a.denom(), 13);

Implementations

impl SmallRational

pub const fn new() -> Self

Creates a SmallRational with value 0.

Examples

#![allow(deprecated)]

use rug::rational::SmallRational;
let r = SmallRational::new();
// Use r as if it were Rational.
assert_eq!(*r.numer(), 0);
assert_eq!(*r.denom(), 1);

pub unsafe fn as_nonreallocating_rational(&mut self) -> &mut Rational

Returns a mutable reference to a Rational number for simple operations that do not need to allocate more space for the numerator or denominator.

Safety

It is undefined behavior to perform operations that reallocate the internal data of the referenced Rational number or to swap it with another number, although it is allowed to swap the numerator and denominator allocations, such as in the reciprocal operation recip_mut.

Some GMP functions swap the allocations of their target operands; calling such functions with the mutable reference returned by this method can lead to undefined behavior.

Examples

#![allow(deprecated)]

use rug::rational::SmallRational;
let mut r = SmallRational::from((-15i32, 47i32));
let num_capacity = r.numer().capacity();
let den_capacity = r.denom().capacity();
// reciprocating this will not require reallocations
unsafe {
    r.as_nonreallocating_rational().recip_mut();
}
assert_eq!(*r, SmallRational::from((-47, 15)));
assert_eq!(r.numer().capacity(), num_capacity);
assert_eq!(r.denom().capacity(), den_capacity);

pub unsafe fn from_canonical<Num: ToSmall, Den: ToSmall>( num: Num, den: Den, ) -> Self

Creates a SmallRational from a numerator and denominator, assuming they are in canonical form.

Safety

This method leads to undefined behavior if den is zero or if num and den have common factors.

Examples

#![allow(deprecated)]

use rug::rational::SmallRational;
let from_unsafe = unsafe { SmallRational::from_canonical(-13, 10) };
// from_safe is canonicalized to the same form as from_unsafe
let from_safe = SmallRational::from((130, -100));
assert_eq!(from_unsafe.numer(), from_safe.numer());
assert_eq!(from_unsafe.denom(), from_safe.denom());

pub unsafe fn assign_canonical<Num: ToSmall, Den: ToSmall>( &mut self, num: Num, den: Den, )

Assigns a numerator and denominator to a SmallRational, assuming they are in canonical form.

Safety

This method leads to undefined behavior if den is zero or negative, or if num and den have common factors.

Examples

#![allow(deprecated)]

use rug::rational::SmallRational;
use rug::Assign;
let mut a = SmallRational::new();
unsafe {
    a.assign_canonical(-13, 10);
}
// b is canonicalized to the same form as a
let mut b = SmallRational::new();
b.assign((130, -100));
assert_eq!(a.numer(), b.numer());
assert_eq!(a.denom(), b.denom());