rug

Struct Rational

pub struct Rational { /* private fields */ }

An arbitrary-precision rational number.

A Rational number is made up of a numerator Integer and denominator Integer. After Rational number functions, the number is always in canonical form, that is the denominator is always greater than zero, and there are no common factors. Zero is stored as 0/1.

Examples

use rug::Rational;
let r = Rational::from((-12, 15));
let recip = Rational::from(r.recip_ref());
assert_eq!(recip, Rational::from((-5, 4)));
assert_eq!(recip, -1.25);
// The numerator and denominator are stored in canonical form.
let (num, den) = r.into_numer_denom();
assert_eq!(num, -4);
assert_eq!(den, 5);

When comparing Rational numbers, it is sometimes convenient to compare to a pair of primitive integers, like in the comparison to Rational::from((-5, 4)) in the example above. Rational::from((-5, 4)) requires memory allocation, which can be an expensive operation. To avoid this allocation, when both the numerator and denominator fit in a primitive integer, we can use MiniRational instead, which does not require heap memory allocation.

use rug::rational::MiniRational;
use rug::Rational;
let r = Rational::from((-12, 15));
let recip = Rational::from(r.recip_ref());
assert_eq!(recip, MiniRational::from((-5, 4)));

The Rational number type supports various functions. Most methods have three versions:

The first method consumes the operand.

The second method has a _mut suffix and mutates the operand.

The third method has a _ref suffix and borrows the operand. The returned item is an incomplete-computation value that can be assigned to a Rational number.

use rug::rational::MiniRational;
use rug::Rational;

// 1. consume the operand
let a = Rational::from((-15, 2));
let abs_a = a.abs();
assert_eq!(abs_a, MiniRational::from((15, 2)));

// 2. mutate the operand
let mut b = Rational::from((-17, 2));
b.abs_mut();
assert_eq!(b, MiniRational::from((17, 2)));

// 3. borrow the operand
let c = Rational::from((-19, 2));
let r = c.abs_ref();
let abs_c = Rational::from(r);
assert_eq!(abs_c, MiniRational::from((19, 2)));
// c was not consumed
assert_eq!(c, MiniRational::from((-19, 2)));

Implementations

impl Rational

pub const ZERO: &'static Rational

Zero.

Examples

use rug::Rational;
assert_eq!(*Rational::ZERO, 0);

pub const ONE: &'static Rational

One.

Examples

use rug::Rational;
assert_eq!(*Rational::ONE, 1);

pub const NEG_ONE: &'static Rational

Negative one (1).

Examples

use rug::Rational;
assert_eq!(*Rational::NEG_ONE, -1);

pub fn new() -> Self

Constructs a new arbitrary-precision Rational number with value 0.

Examples

use rug::Rational;
let r = Rational::new();
assert_eq!(r, 0);

pub const unsafe fn from_raw(raw: mpq_t) -> Self

Creates a Rational number from an initialized GMP rational number.

Safety

The function must not be used to create a constant Rational number, though it can be used to create a static Rational number. This is because constant values are copied on use, leading to undefined behavior when they are dropped.

The value must be initialized as a valid mpq_t.

The mpq_t type can be considered as a kind of pointer, so there can be multiple copies of it. Since this function takes over ownership, no other copies of the passed value should exist.

The numerator and denominator must be in canonical form, as the rest of the library assumes that they are. Most GMP functions leave the rational number in canonical form, but assignment functions do not. Check the GMP documentation for details.

Examples

use core::mem::MaybeUninit;
use gmp_mpfr_sys::gmp;
use rug::rational::MiniRational;
use rug::Rational;
let r = unsafe {
    let mut q = MaybeUninit::uninit();
    gmp::mpq_init(q.as_mut_ptr());
    let mut q = q.assume_init();
    gmp::mpq_set_si(&mut q, -145, 10);
    gmp::mpq_canonicalize(&mut q);
    // q is initialized and unique
    Rational::from_raw(q)
};
assert_eq!(r, MiniRational::from((-145, 10)));
// since r is a Rational now, deallocation is automatic

This can be used to create a static Rational number using MPZ_ROINIT_N to initialize the raw numerator and denominator values. See the GMP documentation for details.

use gmp_mpfr_sys::gmp;
use gmp_mpfr_sys::gmp::{limb_t, mpq_t};
use rug::{Integer, Rational};
const NUMER_LIMBS: [limb_t; 2] = [0, 5];
const DENOM_LIMBS: [limb_t; 1] = [3];
const MPQ: mpq_t = unsafe {
    mpq_t {
        num: gmp::MPZ_ROINIT_N(NUMER_LIMBS.as_ptr().cast_mut(), -2),
        den: gmp::MPZ_ROINIT_N(DENOM_LIMBS.as_ptr().cast_mut(), 1),
    }
};
// Must *not* be const, otherwise it would lead to undefined
// behavior on use, as it would create a copy that is dropped.
static R: Rational = unsafe { Rational::from_raw(MPQ) };
let numer_check =
    -((Integer::from(NUMER_LIMBS[1]) << gmp::NUMB_BITS) + NUMER_LIMBS[0]);
let denom_check = Integer::from(DENOM_LIMBS[0]);
assert_eq!(*R.numer(), numer_check);
assert_eq!(*R.denom(), denom_check);
let check = Rational::from((&numer_check, &denom_check));
assert_eq!(R, check);
assert_eq!(*R.numer(), *check.numer());
assert_eq!(*R.denom(), *check.denom());

pub const fn into_raw(self) -> mpq_t

Converts a Rational number into a GMP rational number.

The returned object should be freed to avoid memory leaks.

Examples

use gmp_mpfr_sys::gmp;
use rug::Rational;
let r = Rational::from((-145, 10));
let mut q = r.into_raw();
unsafe {
    let d = gmp::mpq_get_d(&q);
    assert_eq!(d, -14.5);
    // free object to prevent memory leak
    gmp::mpq_clear(&mut q);
}

pub const fn as_raw(&self) -> *const mpq_t

Returns a pointer to the inner GMP rational number.

The returned pointer will be valid for as long as self is valid.

Examples

use gmp_mpfr_sys::gmp;
use rug::rational::MiniRational;
use rug::Rational;
let r = Rational::from((-145, 10));
let q_ptr = r.as_raw();
unsafe {
    let d = gmp::mpq_get_d(q_ptr);
    assert_eq!(d, -14.5);
}
// r is still valid
assert_eq!(r, MiniRational::from((-145, 10)));

pub fn as_raw_mut(&mut self) -> *mut mpq_t

Returns an unsafe mutable pointer to the inner GMP rational number.

The returned pointer will be valid for as long as self is valid.

Examples

use gmp_mpfr_sys::gmp;
use rug::rational::MiniRational;
use rug::Rational;
let mut r = Rational::from((-145, 10));
let q_ptr = r.as_raw_mut();
unsafe {
    gmp::mpq_inv(q_ptr, q_ptr);
}
assert_eq!(r, MiniRational::from((-10, 145)));

pub fn from_f32(value: f32) -> Option<Self>

Creates a Rational number from an f32 if it is finite, losing no precision.

This conversion can also be performed using

Rational::try_from(value)

value.checked_as::<Rational>()

Examples

use rug::rational::MiniRational;
use rug::Rational;
// -17.125 can be stored exactly as f32
let r = Rational::from_f32(-17.125).unwrap();
assert_eq!(r, MiniRational::from((-17125, 1000)));
let inf = Rational::from_f32(f32::INFINITY);
assert!(inf.is_none());

pub fn from_f64(value: f64) -> Option<Self>

Creates a Rational number from an f64 if it is finite, losing no precision.

This conversion can also be performed using

Rational::try_from(value)

value.checked_as::<Rational>()

Examples

use rug::rational::MiniRational;
use rug::Rational;
// -17.125 can be stored exactly as f64
let r = Rational::from_f64(-17.125).unwrap();
assert_eq!(r, MiniRational::from((-17125, 1000)));
let inf = Rational::from_f64(f64::INFINITY);
assert!(inf.is_none());

pub fn from_str_radix(src: &str, radix: i32) -> Result<Self, ParseRationalError>

Parses a Rational number.

Panics

Panics if radix is less than 2 or greater than 36.

Examples

use rug::rational::MiniRational;
use rug::Rational;
let r1 = Rational::from_str_radix("ff/a", 16).unwrap();
assert_eq!(r1, MiniRational::from((255, 10)));
let r2 = Rational::from_str_radix("+ff0/a0", 16).unwrap();
assert_eq!(r2, MiniRational::from((0xff0, 0xa0)));
assert_eq!(*r2.numer(), 51);
assert_eq!(*r2.denom(), 2);

pub fn parse<S: AsRef<[u8]>>( src: S, ) -> Result<ParseIncomplete, ParseRationalError>

Parses a decimal string slice (&str) or byte slice (&[u8]) into a Rational number.

The following are implemented with the unwrapped returned incomplete-computation value as Src:

Assign<Src> for Rational

From<Src> for Rational

Complete<Completed = Rational> for Src

The string must contain a numerator, and may contain a denominator; the numerator and denominator are separated with a /. The numerator can start with an optional minus or plus sign.

ASCII whitespace is ignored everywhere in the string. Underscores are ignored anywhere except before the first digit of the numerator and between the / and the the first digit of the denominator.

Examples

use rug::rational::MiniRational;
use rug::{Complete, Rational};

assert_eq!(
    Rational::parse("-12/23").unwrap().complete(),
    MiniRational::from((-12, 23))
);
assert_eq!(
    Rational::parse("+ 12 / 23").unwrap().complete(),
    MiniRational::from((12, 23))
);

let invalid = Rational::parse("12/");
assert!(invalid.is_err());

pub fn parse_radix<S: AsRef<[u8]>>( src: S, radix: i32, ) -> Result<ParseIncomplete, ParseRationalError>

Parses a string slice (&str) or byte slice (&[u8]) into a Rational number.

The following are implemented with the unwrapped returned incomplete-computation value as Src:

Assign<Src> for Rational

From<Src> for Rational

Complete<Completed = Rational> for Src

The string must contain a numerator, and may contain a denominator; the numerator and denominator are separated with a /. The numerator can start with an optional minus or plus sign.

ASCII whitespace is ignored everywhere in the string. Underscores are ignored anywhere except before the first digit of the numerator and between the / and the the first digit of the denominator.

If parsing is already done by an external function, the unsafe low-level Integer::assign_bytes_radix_unchecked method can be used in conjunction with the mutate_numer_denom method.

Panics

Panics if radix is less than 2 or greater than 36.

Examples

use rug::rational::MiniRational;
use rug::{Complete, Rational};

let valid1 = Rational::parse_radix("12/23", 4);
assert_eq!(
    valid1.unwrap().complete(),
    MiniRational::from((2 + 4 * 1, 3 + 4 * 2))
);
let valid2 = Rational::parse_radix("12 / yz", 36);
assert_eq!(
    valid2.unwrap().complete(),
    MiniRational::from((2 + 36 * 1, 35 + 36 * 34))
);

let invalid = Rational::parse_radix("12/23", 3);
assert!(invalid.is_err());

If parsing is done externally, low-level code can be used.

use rug::rational::MiniRational;
use rug::Rational;

let num_bytes = &[1, 2];
let den_bytes = &[2, 3];
let radix = 10;
let neg = true;
let mut r = Rational::new();
// SAFETY: radix and the bytes are in the required ranges
r.mutate_numer_denom(|num, den| unsafe {
    num.assign_bytes_radix_unchecked(num_bytes, radix, neg);
    den.assign_bytes_radix_unchecked(den_bytes, radix, false);
});
// -12/23
assert_eq!(r, MiniRational::from((-12, 23)));

pub fn to_f32(&self) -> f32

Converts to an f32, rounding towards zero.

This conversion can also be performed using

(&rational).az::<f32>()

rational.borrow().az::<f32>()

rational.az::<f32>()

Examples

use rug::rational::MiniRational;
use rug::Rational;
let min = Rational::from_f32(f32::MIN).unwrap();
let minus_small = min - MiniRational::from((7, 2)).borrow_excl();
// minus_small is truncated to f32::MIN
assert_eq!(minus_small.to_f32(), f32::MIN);
let times_three_two = minus_small * MiniRational::from((3, 2)).borrow_excl();
// times_three_two is too small
assert_eq!(times_three_two.to_f32(), f32::NEG_INFINITY);

pub fn to_f64(&self) -> f64

Converts to an f64, rounding towards zero.

This conversion can also be performed using

(&rational).az::<f64>()

rational.borrow().az::<f64>()

rational.az::<f64>()

Examples

use rug::rational::MiniRational;
use rug::Rational;

// An `f64` has 53 bits of precision.
let exact = 0x1f_1234_5678_9aff_u64;
let den = 0x1000_u64;
let r = Rational::from((exact, den));
assert_eq!(r.to_f64(), exact as f64 / den as f64);

// large has 56 ones
let large = 0xff_1234_5678_9aff_u64;
// trunc has 53 ones followed by 3 zeros
let trunc = 0xff_1234_5678_9af8_u64;
let j = Rational::from((large, den));
assert_eq!(j.to_f64(), trunc as f64 / den as f64);

let max = Rational::from_f64(f64::MAX).unwrap();
let plus_small = max + MiniRational::from((7, 2)).borrow_excl();
// plus_small is truncated to f64::MAX
assert_eq!(plus_small.to_f64(), f64::MAX);
let times_three_two = plus_small * MiniRational::from((3, 2)).borrow_excl();
// times_three_two is too large
assert_eq!(times_three_two.to_f64(), f64::INFINITY);

pub fn to_string_radix(&self, radix: i32) -> String

Returns a string representation for the specified radix.

Panics

Panics if radix is less than 2 or greater than 36.

Examples

use rug::Rational;
let r1 = Rational::from(0);
assert_eq!(r1.to_string_radix(10), "0");
let r2 = Rational::from((15, 5));
assert_eq!(r2.to_string_radix(10), "3");
let r3 = Rational::from((10, -6));
assert_eq!(r3.to_string_radix(10), "-5/3");
assert_eq!(r3.to_string_radix(5), "-10/3");

pub fn assign_f32(&mut self, val: f32) -> Result<(), ()>

Assigns from an f32 if it is finite, losing no precision.

Examples

use rug::rational::MiniRational;
use rug::Rational;
let mut r = Rational::new();
let ret = r.assign_f32(12.75);
assert!(ret.is_ok());
assert_eq!(r, MiniRational::from((1275, 100)));
let ret = r.assign_f32(f32::NAN);
assert!(ret.is_err());
assert_eq!(r, MiniRational::from((1275, 100)));

pub fn assign_f64(&mut self, val: f64) -> Result<(), ()>

Assigns from an f64 if it is finite, losing no precision.

Examples

use rug::rational::MiniRational;
use rug::Rational;
let mut r = Rational::new();
let ret = r.assign_f64(12.75);
assert!(ret.is_ok());
assert_eq!(r, MiniRational::from((1275, 100)));
let ret = r.assign_f64(1.0 / 0.0);
assert!(ret.is_err());
assert_eq!(r, MiniRational::from((1275, 100)));

pub unsafe fn from_canonical<Num, Den>(num: Num, den: Den) -> Self
where Integer: From<Num> + From<Den>,

Creates a new Rational number from a numerator and denominator without canonicalizing aftwerwards.

Safety

This function is unsafe because it does not canonicalize the Rational number. The caller must ensure that the numerator and denominator are in canonical form, as the rest of the library assumes that they are.

There are a few methods that can be called on Rational numbers that are not in canonical form:

numer and denom, which treat the numerator and denominator separately

assignment methods, which overwrite the previous value and leave the number in canonical form

mutate_numer_denom, which treats the numerator and denominator seprarately, and leaves the number in canoncial form

Examples

use rug::rational::MiniRational;
use rug::Rational;

// -3/5 is in canonical form
let r = unsafe { Rational::from_canonical(-3, 5) };
assert_eq!(r, MiniRational::from((-3, 5)));

pub unsafe fn assign_canonical<Num, Den>(&mut self, num: Num, den: Den)
where Integer: Assign<Num> + Assign<Den>,

Assigns to the numerator and denominator without canonicalizing aftwerwards.

Safety

This function is unsafe because it does not canonicalize the Rational number after the assignment. The caller must ensure that the numerator and denominator are in canonical form, as the rest of the library assumes that they are.

There are a few methods that can be called on Rational numbers that are not in canonical form:

numer and denom, which treat the numerator and denominator separately

assignment methods, which overwrite the previous value and leave the number in canonical form

mutate_numer_denom, which treats the numerator and denominator seprarately, and leaves the number in canoncial form

Examples

use rug::rational::MiniRational;
use rug::Rational;

let mut r = Rational::new();
// -3/5 is in canonical form
unsafe {
    r.assign_canonical(-3, 5);
}
assert_eq!(r, MiniRational::from((-3, 5)));

pub const fn numer(&self) -> &Integer

Borrows the numerator as an Integer.

Examples

use rug::Rational;
let r = Rational::from((12, -20));
// r will be canonicalized to -3/5
assert_eq!(*r.numer(), -3)

pub const fn denom(&self) -> &Integer

Borrows the denominator as an Integer.

Examples

use rug::Rational;
let r = Rational::from((12, -20));
// r will be canonicalized to -3/5
assert_eq!(*r.denom(), 5);

pub fn mutate_numer_denom<F>(&mut self, func: F)
where F: FnOnce(&mut Integer, &mut Integer),

Calls a function with mutable references to the numerator and denominator, then canonicalizes the number.

The denominator must not be zero when the function returns.

Panics

Panics if the denominator is zero when the function returns.

Examples

use rug::Rational;
let mut r = Rational::from((3, 5));
r.mutate_numer_denom(|num, den| {
    // change r from 3/5 to 4/8, which is equal to 1/2
    *num += 1;
    *den += 3;
});
assert_eq!(*r.numer(), 1);
assert_eq!(*r.denom(), 2);

This method does not check that the numerator and denominator are in canonical form before calling func. This means that this method can be used to canonicalize the number after some unsafe methods that do not leave the number in cononical form.

use rug::Rational;
let mut r = Rational::from((3, 5));
unsafe {
    // leave r in non-canonical form
    *r.as_mut_numer_denom_no_canonicalization().0 += 1;
    *r.as_mut_numer_denom_no_canonicalization().1 -= 13;
}
// At this point, r is still not canonical: 4 / -8
assert_eq!(*r.numer(), 4);
assert_eq!(*r.denom(), -8);
r.mutate_numer_denom(|_, _| {});
// Now r is in canonical form: -1 / 2
assert_eq!(*r.numer(), -1);
assert_eq!(*r.denom(), 2);

pub unsafe fn as_mut_numer_denom_no_canonicalization( &mut self, ) -> (&mut Integer, &mut Integer)

Borrows the numerator and denominator mutably without canonicalizing aftwerwards.

Safety

This function is unsafe because it does not canonicalize the Rational number when the borrow ends. The caller must ensure that the numerator and denominator are left in canonical form, as the rest of the library assumes that they are.

There are a few methods that can be called on Rational numbers that are not in canonical form:

numer and denom, which treat the numerator and denominator separately

assignment methods, which overwrite the previous value and leave the number in canonical form

mutate_numer_denom, which treats the numerator and denominator seprarately, and leaves the number in canoncial form

Examples

use rug::rational::MiniRational;
use rug::Rational;

let mut r = Rational::from((3, 5));
{
    let (num, den) = unsafe { r.as_mut_numer_denom_no_canonicalization() };
    // Add one to r by adding den to num. Since num and den
    // are relatively prime, r remains in canonical form.
    *num += &*den;
}
assert_eq!(r, MiniRational::from((8, 5)));

This method can also be used to group some operations before canonicalization. This is usually not beneficial, as early canonicalization usually means subsequent arithmetic operations have less work to do.

use rug::Rational;
let mut r = Rational::from((3, 5));
unsafe {
    // first operation: add 1 to numerator
    *r.as_mut_numer_denom_no_canonicalization().0 += 1;
    // second operation: subtract 13 from denominator
    *r.as_mut_numer_denom_no_canonicalization().1 -= 13;
}
// At this point, r is still not canonical: 4 / -8
assert_eq!(*r.numer(), 4);
assert_eq!(*r.denom(), -8);
r.mutate_numer_denom(|_, _| {});
// Now r is in canonical form: -1 / 2
assert_eq!(*r.numer(), -1);
assert_eq!(*r.denom(), 2);

pub const fn into_numer_denom(self) -> (Integer, Integer)

Converts into numerator and denominator Integer values.

This function reuses the allocated memory and does not allocate any new memory.

Examples

use rug::Rational;
let r = Rational::from((12, -20));
// r will be canonicalized to -3/5
let (num, den) = r.into_numer_denom();
assert_eq!(num, -3);
assert_eq!(den, 5);

pub const fn as_neg(&self) -> BorrowRational<'_>

Borrows a negated copy of the Rational number.

The returned object implements Deref<Target = Rational>.

This method performs a shallow copy and negates it, and negation does not change the allocated data.

Examples

use rug::rational::MiniRational;
use rug::Rational;
let r = Rational::from((7, 11));
let neg_r = r.as_neg();
assert_eq!(*neg_r, MiniRational::from((-7, 11)));
// methods taking &self can be used on the returned object
let reneg_r = neg_r.as_neg();
assert_eq!(*reneg_r, MiniRational::from((7, 11)));
assert_eq!(*reneg_r, r);

pub const fn as_abs(&self) -> BorrowRational<'_>

Borrows an absolute copy of the Rational number.

The returned object implements Deref<Target = Rational>.

This method performs a shallow copy and possibly negates it, and negation does not change the allocated data.

Examples

use rug::rational::MiniRational;
use rug::Rational;
let r = Rational::from((-7, 11));
let abs_r = r.as_abs();
assert_eq!(*abs_r, MiniRational::from((7, 11)));
// methods taking &self can be used on the returned object
let reabs_r = abs_r.as_abs();
assert_eq!(*reabs_r, MiniRational::from((7, 11)));
assert_eq!(*reabs_r, *abs_r);

pub const fn as_recip(&self) -> BorrowRational<'_>

Borrows a reciprocal copy of the Rational number.

The returned object implements Deref<Target = Rational>.

This method performs some shallow copying, swapping numerator and denominator and making sure the sign is in the numerator.

Panics

Panics if the value is zero.

Examples

use rug::rational::MiniRational;
use rug::Rational;
let r = Rational::from((-7, 11));
let recip_r = r.as_recip();
assert_eq!(*recip_r, MiniRational::from((-11, 7)));
// methods taking &self can be used on the returned object
let rerecip_r = recip_r.as_recip();
assert_eq!(*rerecip_r, MiniRational::from((-7, 11)));
assert_eq!(*rerecip_r, r);

pub const fn is_zero(&self) -> bool

Returns true if the number is zero.

Examples

use rug::Rational;
assert!(Rational::from(0).is_zero());
assert!(!(Rational::from((1, 2)).is_zero()));

pub const fn is_positive(&self) -> bool

Returns true if the number is positive and false if the number is zero or negative.

Examples

use rug::Rational;
assert!(Rational::from((1, 2)).is_positive());
assert!(!(Rational::from((-1, 2)).is_positive()));

pub const fn is_negative(&self) -> bool

Returns true if the number is negative and false if the number is zero or positive.

Examples

use rug::Rational;
assert!(Rational::from((-1, 2)).is_negative());
assert!(!(Rational::from((1, 2)).is_negative()));

pub const fn cmp0(&self) -> Ordering

Returns the same result as self.cmp(&0.into()), but is faster.

Examples

use core::cmp::Ordering;
use rug::Rational;
assert_eq!(Rational::from((-5, 7)).cmp0(), Ordering::Less);
assert_eq!(Rational::from(0).cmp0(), Ordering::Equal);
assert_eq!(Rational::from((5, 7)).cmp0(), Ordering::Greater);

pub fn cmp_abs(&self, other: &Self) -> Ordering

Compares the absolute values.

Examples

use core::cmp::Ordering;
use rug::Rational;
let a = Rational::from((-23, 10));
let b = Rational::from((-47, 5));
assert_eq!(a.cmp(&b), Ordering::Greater);
assert_eq!(a.cmp_abs(&b), Ordering::Less);

pub const fn is_integer(&self) -> bool

Returns true if the number is an integer.

Examples

use rug::Rational;
assert!(!(Rational::from((5, 2)).is_integer()));
assert!(Rational::from(3).is_integer());

pub fn sum<'a, I>(values: I) -> SumIncomplete<'a, I>
where I: Iterator<Item = &'a Self>,

Adds a list of Rational values.

The following are implemented with the returned incomplete-computation value as Src:

Assign<Src> for Rational

From<Src> for Rational

Complete<Completed = Rational> for Src

AddAssign<Src> for Rational

Add<Src> for Rational, Add<Rational> for Src

SubAssign<Src> for Rational, SubFrom<Src> for Rational

Sub<Src> for Rational, Sub<Rational> for Src

Examples

use rug::rational::MiniRational;
use rug::{Complete, Rational};

let values = [
    Rational::from((5, 2)),
    Rational::from((-100_000, 7)),
    Rational::from(-4),
];

let sum = Rational::sum(values.iter()).complete();
let expected = MiniRational::from((5 * 7 - 100_000 * 2 - 4 * 14, 14));
assert_eq!(sum, expected);

pub fn dot<'a, I>(values: I) -> DotIncomplete<'a, I>
where I: Iterator<Item = (&'a Self, &'a Self)>,

Finds the dot product of a list of Rational value pairs.

The following are implemented with the returned incomplete-computation value as Src:

Assign<Src> for Rational

From<Src> for Rational

Complete<Completed = Rational> for Src

AddAssign<Src> for Rational

Add<Src> for Rational, Add<Rational> for Src

SubAssign<Src> for Rational, SubFrom<Src> for Rational

Sub<Src> for Rational, Sub<Rational> for Src

Examples

use rug::rational::MiniRational;
use rug::{Complete, Rational};

let a = [Rational::from((270, 7)), Rational::from((-11, 10))];
let b = [Rational::from(7), Rational::from((1, 2))];

let dot = Rational::dot(a.iter().zip(b.iter())).complete();
let expected = MiniRational::from((270 * 20 - 11, 20));
assert_eq!(dot, expected);

pub fn product<'a, I>(values: I) -> ProductIncomplete<'a, I>
where I: Iterator<Item = &'a Self>,

Multiplies a list of Rational values.

The following are implemented with the returned incomplete-computation value as Src:

Assign<Src> for Rational

From<Src> for Rational

Complete<Completed = Rational> for Src

MulAssign<Src> for Rational

Mul<Src> for Rational, Mul<Rational> for Src

Examples

use rug::rational::MiniRational;
use rug::{Complete, Rational};

let values = [
    Rational::from((5, 2)),
    Rational::from((-100_000, 7)),
    Rational::from(-4),
];

let product = Rational::product(values.iter()).complete();
let expected = MiniRational::from((5 * -100_000 * -4, 2 * 7));
assert_eq!(product, expected);

pub fn abs(self) -> Self

Computes the absolute value.

Examples

use rug::rational::MiniRational;
use rug::Rational;
let r = Rational::from((-100, 17));
let abs = r.abs();
assert_eq!(abs, MiniRational::from((100, 17)));

pub fn abs_mut(&mut self)

Computes the absolute value.

Examples

use rug::rational::MiniRational;
use rug::Rational;
let mut r = Rational::from((-100, 17));
r.abs_mut();
assert_eq!(r, MiniRational::from((100, 17)));

pub fn abs_ref(&self) -> AbsIncomplete<'_>

Computes the absolute value.

The following are implemented with the returned incomplete-computation value as Src:

Assign<Src> for Rational

From<Src> for Rational

Complete<Completed = Rational> for Src

Examples

use rug::rational::MiniRational;
use rug::{Complete, Rational};
let r = Rational::from((-100, 17));
let abs = r.abs_ref().complete();
assert_eq!(abs, MiniRational::from((100, 17)));

pub fn signum(self) -> Self

Computes the signum.

0 if the value is zero

1 if the value is positive

1 if the value is negative

Examples

use rug::Rational;
let r = Rational::from((-100, 17));
let signum = r.signum();
assert_eq!(signum, -1);

pub fn signum_mut(&mut self)

Computes the signum.

0 if the value is zero

1 if the value is positive

1 if the value is negative

Examples

use rug::Rational;
let mut r = Rational::from((-100, 17));
r.signum_mut();
assert_eq!(r, -1);

pub fn signum_ref(&self) -> SignumIncomplete<'_>

Computes the signum.

0 if the value is zero

1 if the value is positive

1 if the value is negative

The following are implemented with the returned incomplete-computation value as Src:

Assign<Src> for Integer

Assign<Src> for Rational

From<Src> for Integer

From<Src> for Rational

Complete<Completed = Rational> for Src

Examples

use rug::{Integer, Rational};
let r = Rational::from((-100, 17));
let r_ref = r.signum_ref();
let signum = Integer::from(r_ref);
assert_eq!(signum, -1);

pub fn clamp<Min, Max>(self, min: &Min, max: &Max) -> Self
where Self: PartialOrd<Min> + PartialOrd<Max> + for<'a> Assign<&'a Min> + for<'a> Assign<&'a Max>,

Clamps the value within the specified bounds.

Panics

Panics if the maximum value is less than the minimum value.

Examples

use rug::rational::MiniRational;
use rug::Rational;
let min = MiniRational::from((-3, 2));
let max = MiniRational::from((3, 2));
let min_borrow = &*min.borrow();
let max_borrow = &*max.borrow();
let too_small = Rational::from((-5, 2));
let clamped1 = too_small.clamp(min_borrow, max_borrow);
assert_eq!(clamped1, MiniRational::from((-3, 2)));
let in_range = Rational::from((1, 2));
let clamped2 = in_range.clamp(min_borrow, max_borrow);
assert_eq!(clamped2, MiniRational::from((1, 2)));

pub fn clamp_mut<Min, Max>(&mut self, min: &Min, max: &Max)
where Self: PartialOrd<Min> + PartialOrd<Max> + for<'a> Assign<&'a Min> + for<'a> Assign<&'a Max>,

Clamps the value within the specified bounds.

Panics

Panics if the maximum value is less than the minimum value.

Examples

use rug::rational::MiniRational;
use rug::Rational;
let min = MiniRational::from((-3, 2));
let max = MiniRational::from((3, 2));
let min_borrow = &*min.borrow();
let max_borrow = &*max.borrow();
let mut too_small = Rational::from((-5, 2));
too_small.clamp_mut(min_borrow, max_borrow);
assert_eq!(too_small, MiniRational::from((-3, 2)));
let mut in_range = Rational::from((1, 2));
in_range.clamp_mut(min_borrow, max_borrow);
assert_eq!(in_range, MiniRational::from((1, 2)));

pub fn clamp_ref<'min, 'max, Min, Max>( &self, min: &'min Min, max: &'max Max, ) -> ClampIncomplete<'_, 'min, 'max, Min, Max>
where Self: PartialOrd<Min> + PartialOrd<Max> + for<'a> Assign<&'a Min> + for<'a> Assign<&'a Max>,

Clamps the value within the specified bounds.

The following are implemented with the returned incomplete-computation value as Src:

Assign<Src> for Rational

From<Src> for Rational

Complete<Completed = Rational> for Src

Panics

Panics if the maximum value is less than the minimum value.

Examples

use rug::rational::MiniRational;
use rug::{Assign, Complete, Rational};
let min = MiniRational::from((-3, 2));
let max = MiniRational::from((3, 2));
let min_borrow = &*min.borrow();
let max_borrow = &*max.borrow();
let too_small = Rational::from((-5, 2));
let mut clamped = too_small.clamp_ref(min_borrow, max_borrow).complete();
assert_eq!(clamped, MiniRational::from((-3, 2)));
let in_range = Rational::from((1, 2));
clamped.assign(in_range.clamp_ref(min_borrow, max_borrow));
assert_eq!(clamped, MiniRational::from((1, 2)));

pub fn recip(self) -> Self

Computes the reciprocal.

Panics

Panics if the value is zero.

Examples

use rug::rational::MiniRational;
use rug::Rational;
let r = Rational::from((-100, 17));
let recip = r.recip();
assert_eq!(recip, MiniRational::from((-17, 100)));

pub fn recip_mut(&mut self)

Computes the reciprocal.

This method never reallocates or copies the heap data. It simply swaps the allocated data of the numerator and denominator and makes sure the denominator is stored as positive.

Panics

Panics if the value is zero.

Examples

use rug::rational::MiniRational;
use rug::Rational;
let mut r = Rational::from((-100, 17));
r.recip_mut();
assert_eq!(r, MiniRational::from((-17, 100)));

pub fn recip_ref(&self) -> RecipIncomplete<'_>

Computes the reciprocal.

The following are implemented with the returned incomplete-computation value as Src:

Assign<Src> for Rational

From<Src> for Rational

Complete<Completed = Rational> for Src

Examples

use rug::rational::MiniRational;
use rug::{Complete, Rational};
let r = Rational::from((-100, 17));
assert_eq!(r.recip_ref().complete(), MiniRational::from((-17, 100)));

pub fn trunc(self) -> Self

Rounds the number towards zero.

Examples

use rug::Rational;
// -3.7
let r1 = Rational::from((-37, 10));
let trunc1 = r1.trunc();
assert_eq!(trunc1, -3);
// 3.3
let r2 = Rational::from((33, 10));
let trunc2 = r2.trunc();
assert_eq!(trunc2, 3);

pub fn trunc_mut(&mut self)

Rounds the number towards zero.

Examples

use rug::{Assign, Rational};
// -3.7
let mut r = Rational::from((-37, 10));
r.trunc_mut();
assert_eq!(r, -3);
// 3.3
r.assign((33, 10));
r.trunc_mut();
assert_eq!(r, 3);

pub fn trunc_ref(&self) -> TruncIncomplete<'_>

Rounds the number towards zero.

The following are implemented with the returned incomplete-computation value as Src:

Assign<Src> for Integer

Assign<Src> for Rational

From<Src> for Integer

From<Src> for Rational

Complete<Completed = Rational> for Src

Examples

use rug::{Assign, Integer, Rational};
let mut trunc = Integer::new();
// -3.7
let r1 = Rational::from((-37, 10));
trunc.assign(r1.trunc_ref());
assert_eq!(trunc, -3);
// 3.3
let r2 = Rational::from((33, 10));
trunc.assign(r2.trunc_ref());
assert_eq!(trunc, 3);

pub fn rem_trunc(self) -> Self

Computes the fractional part of the number.

Examples

use rug::rational::MiniRational;
use rug::Rational;
// -100/17 = -5 - 15/17
let r = Rational::from((-100, 17));
let rem = r.rem_trunc();
assert_eq!(rem, MiniRational::from((-15, 17)));

pub fn rem_trunc_mut(&mut self)

Computes the fractional part of the number.

Examples

use rug::rational::MiniRational;
use rug::Rational;
// -100/17 = -5 - 15/17
let mut r = Rational::from((-100, 17));
r.rem_trunc_mut();
assert_eq!(r, MiniRational::from((-15, 17)));

pub fn rem_trunc_ref(&self) -> RemTruncIncomplete<'_>

Computes the fractional part of the number.

The following are implemented with the returned incomplete-computation value as Src:

Assign<Src> for Rational

From<Src> for Rational

Complete<Completed = Rational> for Src

Examples

use rug::rational::MiniRational;
use rug::{Complete, Rational};
// -100/17 = -5 - 15/17
let r = Rational::from((-100, 17));
assert_eq!(r.rem_trunc_ref().complete(), MiniRational::from((-15, 17)));

pub fn fract_trunc(self, trunc: Integer) -> (Self, Integer)

Computes the fractional and truncated parts of the number.

The initial value of trunc is ignored.

Examples

use rug::rational::MiniRational;
use rug::{Integer, Rational};
// -100/17 = -5 - 15/17
let r = Rational::from((-100, 17));
let (fract, trunc) = r.fract_trunc(Integer::new());
assert_eq!(fract, MiniRational::from((-15, 17)));
assert_eq!(trunc, -5);

pub fn fract_trunc_mut(&mut self, trunc: &mut Integer)

Computes the fractional and truncated parts of the number.

The initial value of trunc is ignored.

Examples

use rug::rational::MiniRational;
use rug::{Integer, Rational};
// -100/17 = -5 - 15/17
let mut r = Rational::from((-100, 17));
let mut whole = Integer::new();
r.fract_trunc_mut(&mut whole);
assert_eq!(r, MiniRational::from((-15, 17)));
assert_eq!(whole, -5);

pub fn fract_trunc_ref(&self) -> FractTruncIncomplete<'_>

Computes the fractional and truncated parts of the number.

The following are implemented with the returned incomplete-computation value as Src:

Assign<Src> for (Rational, Integer)

Assign<Src> for (&mut Rational, &mut Integer)

From<Src> for (Rational, Integer)

Complete<Completed = (Rational, Integer)> for Src

Examples

use rug::rational::MiniRational;
use rug::{Assign, Integer, Rational};
// -100/17 = -5 - 15/17
let r = Rational::from((-100, 17));
let r_ref = r.fract_trunc_ref();
let (mut fract, mut trunc) = (Rational::new(), Integer::new());
(&mut fract, &mut trunc).assign(r_ref);
assert_eq!(fract, MiniRational::from((-15, 17)));
assert_eq!(trunc, -5);

pub fn ceil(self) -> Self

Rounds the number upwards (towards plus infinity).

Examples

use rug::Rational;
// -3.7
let r1 = Rational::from((-37, 10));
let ceil1 = r1.ceil();
assert_eq!(ceil1, -3);
// 3.3
let r2 = Rational::from((33, 10));
let ceil2 = r2.ceil();
assert_eq!(ceil2, 4);

pub fn ceil_mut(&mut self)

Rounds the number upwards (towards plus infinity).

Examples

use rug::{Assign, Rational};
// -3.7
let mut r = Rational::from((-37, 10));
r.ceil_mut();
assert_eq!(r, -3);
// 3.3
r.assign((33, 10));
r.ceil_mut();
assert_eq!(r, 4);

pub fn ceil_ref(&self) -> CeilIncomplete<'_>

Rounds the number upwards (towards plus infinity).

The following are implemented with the returned incomplete-computation value as Src:

Assign<Src> for Integer

Assign<Src> for Rational

From<Src> for Integer

From<Src> for Rational

Complete<Completed = Rational> for Src

Examples

use rug::{Assign, Integer, Rational};
let mut ceil = Integer::new();
// -3.7
let r1 = Rational::from((-37, 10));
ceil.assign(r1.ceil_ref());
assert_eq!(ceil, -3);
// 3.3
let r2 = Rational::from((33, 10));
ceil.assign(r2.ceil_ref());
assert_eq!(ceil, 4);

pub fn rem_ceil(self) -> Self

Computes the non-positive remainder after rounding up.

Examples

use rug::rational::MiniRational;
use rug::Rational;
// 100/17 = 6 - 2/17
let r = Rational::from((100, 17));
let rem = r.rem_ceil();
assert_eq!(rem, MiniRational::from((-2, 17)));

pub fn rem_ceil_mut(&mut self)

Computes the non-positive remainder after rounding up.

Examples

use rug::rational::MiniRational;
use rug::Rational;
// 100/17 = 6 - 2/17
let mut r = Rational::from((100, 17));
r.rem_ceil_mut();
assert_eq!(r, MiniRational::from((-2, 17)));

pub fn rem_ceil_ref(&self) -> RemCeilIncomplete<'_>

Computes the non-positive remainder after rounding up.

The following are implemented with the returned incomplete-computation value as Src:

Assign<Src> for Rational

From<Src> for Rational

Complete<Completed = Rational> for Src

Examples

use rug::rational::MiniRational;
use rug::{Complete, Rational};
// 100/17 = 6 - 2/17
let r = Rational::from((100, 17));
assert_eq!(r.rem_ceil_ref().complete(), MiniRational::from((-2, 17)));

pub fn fract_ceil(self, ceil: Integer) -> (Self, Integer)

Computes the fractional and ceil parts of the number.

The fractional part cannot greater than zero.

The initial value of ceil is ignored.

Examples

use rug::rational::MiniRational;
use rug::{Integer, Rational};
// 100/17 = 6 - 2/17
let r = Rational::from((100, 17));
let (fract, ceil) = r.fract_ceil(Integer::new());
assert_eq!(fract, MiniRational::from((-2, 17)));
assert_eq!(ceil, 6);

pub fn fract_ceil_mut(&mut self, ceil: &mut Integer)

Computes the fractional and ceil parts of the number.

The fractional part cannot be greater than zero.

The initial value of ceil is ignored.

Examples

use rug::rational::MiniRational;
use rug::{Integer, Rational};
// 100/17 = 6 - 2/17
let mut r = Rational::from((100, 17));
let mut ceil = Integer::new();
r.fract_ceil_mut(&mut ceil);
assert_eq!(r, MiniRational::from((-2, 17)));
assert_eq!(ceil, 6);

pub fn fract_ceil_ref(&self) -> FractCeilIncomplete<'_>

Computes the fractional and ceil parts of the number.

The fractional part cannot be greater than zero.

The following are implemented with the returned incomplete-computation value as Src:

Assign<Src> for (Rational, Integer)

Assign<Src> for (&mut Rational, &mut Integer)

From<Src> for (Rational, Integer)

Complete<Completed = (Rational, Integer)> for Src

Examples

use rug::rational::MiniRational;
use rug::{Assign, Integer, Rational};
// 100/17 = 6 - 2/17
let r = Rational::from((100, 17));
let r_ref = r.fract_ceil_ref();
let (mut fract, mut ceil) = (Rational::new(), Integer::new());
(&mut fract, &mut ceil).assign(r_ref);
assert_eq!(fract, MiniRational::from((-2, 17)));
assert_eq!(ceil, 6);

pub fn floor(self) -> Self

Rounds the number downwards (towards minus infinity).

Examples

use rug::Rational;
// -3.7
let r1 = Rational::from((-37, 10));
let floor1 = r1.floor();
assert_eq!(floor1, -4);
// 3.3
let r2 = Rational::from((33, 10));
let floor2 = r2.floor();
assert_eq!(floor2, 3);

pub fn floor_mut(&mut self)

Rounds the number downwards (towards minus infinity).

use rug::{Assign, Rational};
// -3.7
let mut r = Rational::from((-37, 10));
r.floor_mut();
assert_eq!(r, -4);
// 3.3
r.assign((33, 10));
r.floor_mut();
assert_eq!(r, 3);

pub fn floor_ref(&self) -> FloorIncomplete<'_>

Rounds the number downwards (towards minus infinity).

The following are implemented with the returned incomplete-computation value as Src:

Assign<Src> for Integer

Assign<Src> for Rational

From<Src> for Integer

From<Src> for Rational

Complete<Completed = Rational> for Src

Examples

use rug::{Assign, Integer, Rational};
let mut floor = Integer::new();
// -3.7
let r1 = Rational::from((-37, 10));
floor.assign(r1.floor_ref());
assert_eq!(floor, -4);
// 3.3
let r2 = Rational::from((33, 10));
floor.assign(r2.floor_ref());
assert_eq!(floor, 3);

pub fn rem_floor(self) -> Self

Computes the non-negative remainder after rounding down.

Examples

use rug::rational::MiniRational;
use rug::Rational;
// -100/17 = -6 + 2/17
let r = Rational::from((-100, 17));
let rem = r.rem_floor();
assert_eq!(rem, MiniRational::from((2, 17)));

pub fn rem_floor_mut(&mut self)

Computes the non-negative remainder after rounding down.

Examples

use rug::rational::MiniRational;
use rug::Rational;
// -100/17 = -6 + 2/17
let mut r = Rational::from((-100, 17));
r.rem_floor_mut();
assert_eq!(r, MiniRational::from((2, 17)));

pub fn rem_floor_ref(&self) -> RemFloorIncomplete<'_>

Computes the non-negative remainder after rounding down.

The following are implemented with the returned incomplete-computation value as Src:

Assign<Src> for Rational

From<Src> for Rational

Complete<Completed = Rational> for Src

Examples

use rug::rational::MiniRational;
use rug::{Complete, Rational};
// -100/17 = -6 + 2/17
let r = Rational::from((-100, 17));
assert_eq!(r.rem_floor_ref().complete(), MiniRational::from((2, 17)));

pub fn fract_floor(self, floor: Integer) -> (Self, Integer)

Computes the fractional and floor parts of the number.

The fractional part cannot be negative.

The initial value of floor is ignored.

Examples

use rug::rational::MiniRational;
use rug::{Integer, Rational};
// -100/17 = -6 + 2/17
let r = Rational::from((-100, 17));
let (fract, floor) = r.fract_floor(Integer::new());
assert_eq!(fract, MiniRational::from((2, 17)));
assert_eq!(floor, -6);

pub fn fract_floor_mut(&mut self, floor: &mut Integer)

Computes the fractional and floor parts of the number.

The fractional part cannot be negative.

The initial value of floor is ignored.

Examples

use rug::rational::MiniRational;
use rug::{Integer, Rational};
// -100/17 = -6 + 2/17
let mut r = Rational::from((-100, 17));
let mut floor = Integer::new();
r.fract_floor_mut(&mut floor);
assert_eq!(r, MiniRational::from((2, 17)));
assert_eq!(floor, -6);

pub fn fract_floor_ref(&self) -> FractFloorIncomplete<'_>

Computes the fractional and floor parts of the number.

The fractional part cannot be negative.

The following are implemented with the returned incomplete-computation value as Src:

Assign<Src> for (Rational, Integer)

Assign<Src> for (&mut Rational, &mut Integer)

From<Src> for (Rational, Integer)

Complete<Completed = (Rational, Integer)> for Src

Examples

use rug::rational::MiniRational;
use rug::{Assign, Integer, Rational};
// -100/17 = -6 + 2/17
let r = Rational::from((-100, 17));
let r_ref = r.fract_floor_ref();
let (mut fract, mut floor) = (Rational::new(), Integer::new());
(&mut fract, &mut floor).assign(r_ref);
assert_eq!(fract, MiniRational::from((2, 17)));
assert_eq!(floor, -6);

pub fn round(self) -> Self

Rounds the number to the nearest integer.

When the number lies exactly between two integers, it is rounded away from zero.

Examples

use rug::Rational;
// -3.5
let r1 = Rational::from((-35, 10));
let round1 = r1.round();
assert_eq!(round1, -4);
// 3.7
let r2 = Rational::from((37, 10));
let round2 = r2.round();
assert_eq!(round2, 4);

pub fn round_mut(&mut self)

Rounds the number to the nearest integer.

When the number lies exactly between two integers, it is rounded away from zero.

Examples

use rug::{Assign, Rational};
// -3.5
let mut r = Rational::from((-35, 10));
r.round_mut();
assert_eq!(r, -4);
// 3.7
r.assign((37, 10));
r.round_mut();
assert_eq!(r, 4);

pub fn round_ref(&self) -> RoundIncomplete<'_>

Rounds the number to the nearest integer.

When the number lies exactly between two integers, it is rounded away from zero.

The following are implemented with the returned incomplete-computation value as Src:

Assign<Src> for Integer`

Assign<Src> for Rational

From<Src> for Integer

From<Src> for Rational

Complete<Completed = Rational> for Src

Examples

use rug::{Assign, Integer, Rational};
let mut round = Integer::new();
// -3.5
let r1 = Rational::from((-35, 10));
round.assign(r1.round_ref());
assert_eq!(round, -4);
// 3.7
let r2 = Rational::from((37, 10));
round.assign(r2.round_ref());
assert_eq!(round, 4);

pub fn rem_round(self) -> Self

Computes the remainder after rounding to the nearest integer.

Examples

use rug::rational::MiniRational;
use rug::Rational;
// -3.5 = -4 + 0.5 = -4 + 1/2
let r1 = Rational::from((-35, 10));
let rem1 = r1.rem_round();
assert_eq!(rem1, MiniRational::from((1, 2)));
// 3.7 = 4 - 0.3 = 4 - 3/10
let r2 = Rational::from((37, 10));
let rem2 = r2.rem_round();
assert_eq!(rem2, MiniRational::from((-3, 10)));

pub fn rem_round_mut(&mut self)

Computes the remainder after rounding to the nearest integer.

Examples

use rug::rational::MiniRational;
use rug::Rational;
// -3.5 = -4 + 0.5 = -4 + 1/2
let mut r1 = Rational::from((-35, 10));
r1.rem_round_mut();
assert_eq!(r1, MiniRational::from((1, 2)));
// 3.7 = 4 - 0.3 = 4 - 3/10
let mut r2 = Rational::from((37, 10));
r2.rem_round_mut();
assert_eq!(r2, MiniRational::from((-3, 10)));

pub fn rem_round_ref(&self) -> RemRoundIncomplete<'_>

Computes the remainder after rounding to the nearest integer.

The following are implemented with the returned incomplete-computation value as Src:

Assign<Src> for Rational

From<Src> for Rational

Complete<Completed = Rational> for Src

Examples

use rug::rational::MiniRational;
use rug::{Assign, Complete, Rational};
// -3.5 = -4 + 0.5 = -4 + 1/2
let r1 = Rational::from((-35, 10));
let mut rem = r1.rem_round_ref().complete();
assert_eq!(rem, MiniRational::from((1, 2)));
// 3.7 = 4 - 0.3 = 4 - 3/10
let r2 = Rational::from((37, 10));
rem.assign(r2.rem_round_ref());
assert_eq!(rem, MiniRational::from((-3, 10)));

pub fn fract_round(self, round: Integer) -> (Self, Integer)

Computes the fractional and rounded parts of the number.

The fractional part is positive when the number is rounded down and negative when the number is rounded up. When the number lies exactly between two integers, it is rounded away from zero.

The initial value of round is ignored.

Examples

use rug::rational::MiniRational;
use rug::{Integer, Rational};
// -3.5 = -4 + 0.5 = -4 + 1/2
let r1 = Rational::from((-35, 10));
let (fract1, round1) = r1.fract_round(Integer::new());
assert_eq!(fract1, MiniRational::from((1, 2)));
assert_eq!(round1, -4);
// 3.7 = 4 - 0.3 = 4 - 3/10
let r2 = Rational::from((37, 10));
let (fract2, round2) = r2.fract_round(Integer::new());
assert_eq!(fract2, MiniRational::from((-3, 10)));
assert_eq!(round2, 4);

pub fn fract_round_mut(&mut self, round: &mut Integer)

Computes the fractional and round parts of the number.

The fractional part is positive when the number is rounded down and negative when the number is rounded up. When the number lies exactly between two integers, it is rounded away from zero.

The initial value of round is ignored.

Examples

use rug::rational::MiniRational;
use rug::{Integer, Rational};
// -3.5 = -4 + 0.5 = -4 + 1/2
let mut r1 = Rational::from((-35, 10));
let mut round1 = Integer::new();
r1.fract_round_mut(&mut round1);
assert_eq!(r1, MiniRational::from((1, 2)));
assert_eq!(round1, -4);
// 3.7 = 4 - 0.3 = 4 - 3/10
let mut r2 = Rational::from((37, 10));
let mut round2 = Integer::new();
r2.fract_round_mut(&mut round2);
assert_eq!(r2, MiniRational::from((-3, 10)));
assert_eq!(round2, 4);

pub fn fract_round_ref(&self) -> FractRoundIncomplete<'_>

Computes the fractional and round parts of the number.

The fractional part is positive when the number is rounded down and negative when the number is rounded up. When the number lies exactly between two integers, it is rounded away from zero.

The following are implemented with the returned incomplete-computation value as Src:

Assign<Src> for (Rational, Integer)

Assign<Src> for (&mut Rational, &mut Integer)

From<Src> for (Rational, Integer)

Complete<Completed = (Rational, Integer)> for Src

Examples

use rug::rational::MiniRational;
use rug::{Assign, Integer, Rational};
// -3.5 = -4 + 0.5 = -4 + 1/2
let r1 = Rational::from((-35, 10));
let r_ref1 = r1.fract_round_ref();
let (mut fract1, mut round1) = (Rational::new(), Integer::new());
(&mut fract1, &mut round1).assign(r_ref1);
assert_eq!(fract1, MiniRational::from((1, 2)));
assert_eq!(round1, -4);
// 3.7 = 4 - 0.3 = 4 - 3/10
let r2 = Rational::from((37, 10));
let r_ref2 = r2.fract_round_ref();
let (mut fract2, mut round2) = (Rational::new(), Integer::new());
(&mut fract2, &mut round2).assign(r_ref2);
assert_eq!(fract2, MiniRational::from((-3, 10)));
assert_eq!(round2, 4);

pub fn square(self) -> Self

Computes the square.

This method cannot be replaced by a multiplication using the * operator: r * r and r * &r are both errors.

Examples

use rug::rational::MiniRational;
use rug::Rational;
let r = Rational::from((-13, 2));
let square = r.square();
assert_eq!(square, MiniRational::from((169, 4)));

pub fn square_mut(&mut self)

Computes the square.

This method cannot be replaced by a compound multiplication and assignment using the *= operataor: r *= r; and r *= &r; are both errors.

Examples

use rug::rational::MiniRational;
use rug::Rational;
let mut r = Rational::from((-13, 2));
r.square_mut();
assert_eq!(r, MiniRational::from((169, 4)));

pub fn square_ref(&self) -> MulIncomplete<'_>

Computes the square.

The following are implemented with the returned incomplete-computation value as Src:

Assign<Src> for Rational

From<Src> for Rational

Complete<Completed = Rational> for Src

r.square_ref() produces the exact same result as &r * &r.

Examples

use rug::rational::MiniRational;
use rug::{Complete, Rational};
let r = Rational::from((-13, 2));
assert_eq!(r.square_ref().complete(), MiniRational::from((169, 4)));